requirements.txt
torch
torchvision
ipykernel
ipywidgets
transformers>=4.50.0
scikit-learn
nltk
pycocotools
poetry
python-dotenv
wandb
git+https://github.com/openai/CLIP.git
